+++ backend/ModernWMS.WMS/Services/GoodsOwner/GoodsownerService.cs
@@ -15,7 +15,7 @@ namespace ModernWMS.WMS.Services
     /// <summary>
     /// Goods owner Service
     /// </summary>
-    public class GoodsownerService : BaseService<GoodsownerEntity>, IGoodsownerService
+    public class GoodsOwnerService : BaseService<GoodsownerEntity>, IGoodsownerService
     {
         #region Args
         /// <summary>
@@ -35,7 +35,7 @@ namespace ModernWMS.WMS.Services
         /// </summary>
         /// <param name="dBContext">The DBContext</param>
         /// <param name="stringLocalizer">Localization</param>
-        public GoodsownerService(
+        public GoodsOwnerService(
             SqlDBContext dBContext
           , IStringLocalizer<Core.MultiLanguage> stringLocalizer
             )
@@ -91,9 +91,9 @@ namespace ModernWMS.WMS.Services
         /// </summary>
         /// <param name="id">id</param>
         /// <returns></returns>
-        public async Task<GoodsownerViewModel> GetAsync(int id)
+        public async Task<GoodsownerViewModel> GetAsync(int id, CurrentUser currentUser)
         {
-            var entity = await _dBContext.GetDbSet<GoodsownerEntity>().AsNoTracking().FirstOrDefaultAsync(t => t.id.Equals(id));
+            var entity = await _dBContext.GetDbSet<GoodsownerEntity>().AsNoTracking().FirstOrDefaultAsync(t => t.id.Equals(id) && t.tenant_id == currentUser.tenant_id);
             if (entity != null)
             {
                 return entity.Adapt<GoodsownerViewModel>();
@@ -138,10 +138,10 @@ namespace ModernWMS.WMS.Services
         /// </summary>
         /// <param name="viewModel">args</param>
         /// <returns></returns>
-        public async Task<(bool flag, string msg)> UpdateAsync(GoodsownerViewModel viewModel)
+        public async Task<(bool flag, string msg)> UpdateAsync(GoodsownerViewModel viewModel, CurrentUser currentUser)
         {
             var DbSet = _dBContext.GetDbSet<GoodsownerEntity>();
-            var entity = await DbSet.FirstOrDefaultAsync(t => t.id.Equals(viewModel.id));
+            var entity = await DbSet.FirstOrDefaultAsync(t => t.id.Equals(viewModel.id) && t.tenant_id == currentUser.tenant_id);
             if (entity == null)
             {
                 return (false, _stringLocalizer["not_exists_entity"]);
@@ -172,9 +172,9 @@ namespace ModernWMS.WMS.Services
         /// </summary>
         /// <param name="id">id</param>
         /// <returns></returns>
-        public async Task<(bool flag, string msg)> DeleteAsync(int id)
+        public async Task<(bool flag, string msg)> DeleteAsync(int id, CurrentUser currentUser)
         {
-            var qty = await _dBContext.GetDbSet<GoodsownerEntity>().Where(t => t.id.Equals(id)).ExecuteDeleteAsync();
+            var qty = await _dBContext.GetDbSet<GoodsownerEntity>().Where(t => t.id.Equals(id) && t.tenant_id == currentUser.tenant_id).ExecuteDeleteAsync();
             if (qty > 0)
             {
                 return (true, _stringLocalizer["delete_success"]);
