+++ backend/ModernWMS.WMS/Services/supplier/SupplierService.cs
@@ -99,10 +99,10 @@ namespace ModernWMS.WMS.Services
          /// Get a record by id
          /// </summary>
          /// <returns></returns>
-         public async Task<SupplierViewModel> GetAsync(int id)
+         public async Task<SupplierViewModel> GetAsync(int id, CurrentUser currentUser)
          {
              var DbSet = _dBContext.GetDbSet<SupplierEntity>();
-             var entity = await DbSet.AsNoTracking().FirstOrDefaultAsync(t=>t.id.Equals(id));
+             var entity = await DbSet.AsNoTracking().FirstOrDefaultAsync(t=>t.id.Equals(id) && t.tenant_id == currentUser.tenant_id);
              if (entity == null)
              {
                  return null;
@@ -115,9 +115,13 @@ namespace ModernWMS.WMS.Services
         /// <param name="viewModel">viewmodel</param>
         /// <param name="currentUser">current user</param>
         /// <returns></returns>
-        public async Task<(int id, string msg)> AddAsync(SupplierViewModel viewModel, CurrentUser currentUser) 
+        public async Task<(int id, string msg)> AddAsync(SupplierViewModel viewModel, CurrentUser currentUser)
         {
             var DbSet = _dBContext.GetDbSet<SupplierEntity>();
+            if (await DbSet.AnyAsync(t => t.supplier_name == viewModel.supplier_name && t.tenant_id == currentUser.tenant_id))
+            {
+                return (0, string.Format(_stringLocalizer["exists_entity"], _stringLocalizer["supplier_name"], viewModel.supplier_name));
+            }
             var entity = viewModel.Adapt<SupplierEntity>();
              entity.id = 0;
              entity.create_time = DateTime.Now;
@@ -125,10 +129,6 @@ namespace ModernWMS.WMS.Services
              entity.last_update_time = DateTime.Now;
              entity.tenant_id = currentUser.tenant_id;
              await DbSet.AddAsync(entity);
-            if (await DbSet.AnyAsync(t => t.supplier_name == viewModel.supplier_name && t.tenant_id == currentUser.tenant_id))
-            {
-                return (0, string.Format(_stringLocalizer["exists_entity"], _stringLocalizer["supplier_name"], viewModel.supplier_name));
-            }
             await _dBContext.SaveChangesAsync();
              if (entity.id > 0)
              {
@@ -152,7 +152,7 @@ namespace ModernWMS.WMS.Services
             {
                return(false, string.Format(_stringLocalizer["exists_entity"], _stringLocalizer["supplier_name"], viewModel.supplier_name));
             }
-            var entity = await DbSet.FirstOrDefaultAsync(t => t.id.Equals(viewModel.id));
+            var entity = await DbSet.FirstOrDefaultAsync(t => t.id.Equals(viewModel.id) && t.tenant_id == currentUser.tenant_id);
              if (entity == null)
              {
                  return (false,_stringLocalizer[ "not_exists_entity"]);
@@ -181,9 +181,9 @@ namespace ModernWMS.WMS.Services
          /// </summary>
          /// <param name="id">id</param>
          /// <returns></returns>
-         public async Task<(bool flag, string msg)> DeleteAsync(int id)
+         public async Task<(bool flag, string msg)> DeleteAsync(int id, CurrentUser currentUser)
          {
-             var qty = await _dBContext.GetDbSet<SupplierEntity>().Where(t => t.id.Equals(id)).ExecuteDeleteAsync();
+             var qty = await _dBContext.GetDbSet<SupplierEntity>().Where(t => t.id.Equals(id) && t.tenant_id == currentUser.tenant_id).ExecuteDeleteAsync();
              if (qty > 0)
              {
                  return (true, _stringLocalizer["delete_success"]);
